MyTelegramBot/requirements.txt
python_telegram_bot==5.3.0
requests==2.13.0